fix: Corrected typo in ICore.hpp...

+++ ICore.hpp
@@ -62,5 +62,5 @@ public:
 
     // This tells the core about the score of a game session that was just finished (i.e. the player got a game over or something like that)
     // Note: This DOES NOT mean that the core should go back to the menu or do anything apart from adding the given score to the score list
-    virtual void addNewScore(std::uint64_t score) = 0;
+    virtual void addNewScore(std::uint32_t score) = 0;
 };
